Updating JPSInstallationDao to use variant.id instead of variantID
Motivation:
When the migrator is run it will update the installation table and
modify the variant_id field to refer to the variant tables id field
instead of variantID field. This was needed to joins to work correctly.

We should also update the quieries to use the variant.id field instead
of the variant.variantId field when joining from the installation table
to the variant table.

Modifications:
Updated JPSInstalltionDao and changed the queries that join the
installation table and the variant table.

Also updated the calls to to use the variant id instead of the variantID
field where applicable.