Introducing a Mysql5BitBooleanDialect.
Motivation:
To overcome the issue of "Found: bit, expected: boolean" for mysql we
are adding a custom mysql dialect for Hibernate. This allows a bit to
used in the database an interpreted as a boolean type.

Modifications:
Apart from adding the custom dialect itself a Hibernate DialectResolver
has also been added. There are two version of this resolver, one for
ups-as7, and one for ups-wildfly as they use different version of
Hibernate and there are differences (packaged names and the interface
itself)

Result:
It is not possible to migrate from an earlier version of UPS running on
Mysql.