Load only enabled installations when checking if an installation is authroized for a variant
Loading the installation into memory and checking whether it is enabled was broken since several
installations may exist for the same variant and device token. Therefore querying for the installation
and returning just one of them may return one who is disabled even though an enabled counterpart exists.

Instead when querying the database for the installation, explicitly include only the enabled ones.