Add an option to overwrite existing ducments during document upload
Sometimes we don't want to create a new document if an older version of it exists,
but rather to overwrite it. In this case, we generate a file name without the timestamp
field, making it so two files files representing the same content (in different points in time)
have the same name.