Adding gcm token validation
From gcm docs:

> When you send a notification to a registration id which is expired, for the 1st time the message(notification) will be delivered but you will get a new registration id with the name canonical id. Which mean, the registration id you sent the message to has been changed to this canonical id, so change it on your server side as well.

I submitted a gcm token in aerogear and used it for some time. after that, the token was expired, and it was removed from aerogear database but not removed from database and not replaced with canonical id.
I checked aerogear code and the feature was not implemented in the sender file:
`org.jboss.aerogear.unifiedpush.message.sender.GCMPushNotificationSender.java`

I was wondering if you could add this feature in the cleanupInvalidRegistrationIDsForVariant function of the mentioned file so the gcm tokens that must be replaced wont be removed.