Replacing AeroGearLogger with slf4j.
Motivation:
To allow for more flexibility when it comes to the logging
implementation used in UPS we want to introduce slf4j. Using this
abstraction the choice of logging framework can be made at deployment
time.

Modifications:
Removed AeroGearLogger and introduced logging using the slf4j API.

Result:
All test pass and server logs are still produces as before.

JIRA:
https://issues.jboss.org/browse/AGPUSH-1086